Add files via upload
The main aim of this project is to develop an android application that is used to control an LED located anywhere across the world.
An  Android  Application  to  control  an  Light  Emitting  Diode(LED)  from  anywhere across the world through internet by using MIT app inventor 2. The Internet connectivity for the LED is enabled by using Arduino uno and Ethernet Shield boards.  A  ‘C’  language  based  code  is  dumped  into  the  Arduino  board  using  Arduino 
interface